PF-30 Fix for skipping question 1
